refactor(types): unify duplicate types to shared package (#36, #48)
- Unify Rol to shared across backend (#36)
- Unify CitaItemInput and CitaCreateInput to shared (#36)
- Align backend Public* types to derive from shared (#48)
- PublicCita = Omit<Cita, 'servicios'>, PublicServicio with documented overrides

Closes #36, Closes #48
