+++ backend/src/domain/logic/citas.items.ts
@@ -1,7 +1,6 @@
-export type CitaItemInput = {
-  servicio_id: string;
-  tipo: 'comprado' | 'canjeado';
-};
+import type { CitaItemInput } from '@fidelity-card/shared';
+
+export type { CitaItemInput };
 
 export type ServicioMaster = {
   id: string;
