+++ backend/src/domain/types/auth.ts
@@ -1,4 +1,5 @@
-export type Rol = 'admin' | 'clienta';
+export type { Rol } from '@fidelity-card/shared';
+export type { Profile as PublicProfile } from '@fidelity-card/shared';
 
 export type SignupBody = {
   email: string;
@@ -16,16 +17,5 @@ export type SigninBody = {
 export type PublicUser = {
   id: string;
   email: string;
-  created_at: string;
-};
-
-export type PublicProfile = {
-  id: string;
-  nombre: string;
-  apellido: string;
-  telefono: string;
-  email: string;
-  rol: Rol;
-  puntos: number;
   created_at: string;
 };
