+++ backend/src/domain/types/referidos.ts
@@ -1,7 +1,1 @@
-export type PublicReferido = {
-  id: string;
-  referente_id: string;
-  referida_id: string;
-  puntos_ganados: number;
-  fecha: string;
-};
+export type { Referido as PublicReferido } from '@fidelity-card/shared';
